elimu space v2/requirements.txt
Django==4.2.7
djongo==1.3.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
pymongo==3.12.3
pillow==10.1.0
python-decouple==3.8
sqlparse==0.2.4

